refactor: move password-token to authentication
