+++ src/apps/backend/modules/authentication/internals/password-reset-token/store/password-reset-token-db.ts
