+++ src/apps/backend/modules/authentication/internals/password-reset-token/store/password-reset-token-repository.ts
@@ -1,4 +1,4 @@
-import { ApplicationRepository } from '../../../application';
+import { ApplicationRepository } from '../../../../application';
 
 import {
   PasswordResetTokenDB,
