+++ src/apps/backend/modules/authentication/rest-api/authentication-serializer.ts
@@ -1,4 +1,4 @@
-import { AccessToken } from '../types';
+import { AccessToken, PasswordResetToken } from '../types';
 
 export const serializeAccessTokenAsJSON = (
   accessToken: AccessToken
@@ -6,4 +6,15 @@ export const serializeAccessTokenAsJSON = (
   accountId: accessToken.accountId,
   expiresAt: accessToken.expiresAt.toUTCString(),
   token: accessToken.token,
+});
+
+export const serializePasswordResetTokenAsJSON = (
+  passwordResetToken: PasswordResetToken
+): unknown => ({
+  id: passwordResetToken.id,
+  account: passwordResetToken.account,
+  expiresAt: passwordResetToken.expiresAt,
+  token: passwordResetToken.token,
+  isUsed: passwordResetToken.isUsed,
+  isExpired: passwordResetToken.isExpired,
 });
