+++ src/apps/backend/modules/communication/index.ts
@@ -1,2 +1,3 @@
 export { default as EmailService } from './email-service';
 export { default as SMSService } from './sms-service';
+export * from './types';
