--- src/apps/backend/modules/password-reset-token/index.ts
@@ -1,3 +1,0 @@
-export { default as PasswordResetTokenServer } from './rest-api/password-reset-token-server';
-export { default as PasswordResetTokenService } from './password-reset-token-service';
-export * from './types';
