--- src/apps/backend/modules/password-reset-token/rest-api/password-reset-token-serializer.ts
@@ -1,12 +1,0 @@
-import { PasswordResetToken } from '../types';
-
-export const serializePasswordResetTokenAsJSON = (
-  passwordResetToken: PasswordResetToken
-): unknown => ({
-  id: passwordResetToken.id,
-  account: passwordResetToken.account,
-  expiresAt: passwordResetToken.expiresAt,
-  token: passwordResetToken.token,
-  isUsed: passwordResetToken.isUsed,
-  isExpired: passwordResetToken.isExpired,
-});
